requirements.txt
Django>=5.1,<5.2
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
PyPDF2==3.0.1
requests==2.32.3
psycopg2-binary==2.9.9
